+++ src/components/ProgressBar/index.js
@@ -1,0 +1,4 @@
+import { ProgressBar } from 'react-bootstrap';
+
+export default ProgressBar;
+export { ProgressBar };