+++ src/components/Spinner/index.js
@@ -1,0 +1,4 @@
+import { Spinner } from 'react-bootstrap';
+
+export default Spinner;
+export { Spinner };